+++ tests/testthat/test-pptx-info.R
@@ -261,6 +261,12 @@ test_that("plot layout properties (part 2)", {
     "Showing current slide's layout: \"Title and Content\"",
     fixed = TRUE
   )
+  # issue 666: specify which slide's layout to plot by index
+  expect_message(
+    plot_layout_properties(x, slide_idx = 1),
+    "Showing layout of slide 1: \"Title and Content\"",
+    fixed = TRUE
+  )
   expect_no_error(plot_layout_properties(x, "Title and Content", legend = TRUE))
 })